Move Checkpoint.Policy to primitive
