+++ lib/primitive/cardano-wallet-primitive.cabal
@@ -105,6 +105,7 @@ library
     Cardano.Wallet.Primitive.Types.Block
     Cardano.Wallet.Primitive.Types.Block.Gen
     Cardano.Wallet.Primitive.Types.BlockSummary
+    Cardano.Wallet.Primitive.Types.Checkpoints.Policy
     Cardano.Wallet.Primitive.Types.Coin
     Cardano.Wallet.Primitive.Types.Coin.Gen
     Cardano.Wallet.Primitive.Types.DecentralizationLevel
